`ApiGatewayProxyHandler` -> `ApiGatewayProxyHandlerV2`
