+++ lambda-http4s/src/main/scala/feral/lambda/http4s/ApiGatewayProxyHandler.scala
@@ -34,66 +34,17 @@ import org.http4s.headers.Cookie
 import org.http4s.headers.`Set-Cookie`
 
 object ApiGatewayProxyHandler {
-  def apply[F[_]: Concurrent: ApiGatewayProxyV2Invocation](
+  @deprecated("Use ApiGatewayProxyHandlerV2", "0.3.0")
+  def apply[F[_]: Concurrent: ApiGatewayProxyInvocationV2](
       routes: HttpRoutes[F]): F[Option[ApiGatewayProxyStructuredResultV2]] = httpRoutes(routes)
 
-  def httpRoutes[F[_]: Concurrent: ApiGatewayProxyV2Invocation](
-      routes: HttpRoutes[F]): F[Option[ApiGatewayProxyStructuredResultV2]] = httpApp(
-    routes.orNotFound)
+  @deprecated("Use ApiGatewayProxyHandlerV2", "0.3.0")
+  def httpRoutes[F[_]: Concurrent: ApiGatewayProxyInvocationV2](
+      routes: HttpRoutes[F]): F[Option[ApiGatewayProxyStructuredResultV2]] =
+    ApiGatewayProxyHandlerV2.httpRoutes(app)
 
-  def httpApp[F[_]: Concurrent: ApiGatewayProxyV2Invocation](
+  @deprecated("Use ApiGatewayProxyHandlerV2", "0.3.0")
+  def httpApp[F[_]: Concurrent: ApiGatewayProxyInvocationV2](
       app: HttpApp[F]): F[Option[ApiGatewayProxyStructuredResultV2]] =
-    for {
-      event <- Invocation.event
-      request <- decodeEvent(event)
-      response <- app(request)
-      isBase64Encoded = !response.charset.contains(Charset.`UTF-8`)
-      responseBody <- response
-        .body
-        .through(
-          if (isBase64Encoded) fs2.text.base64.encode else fs2.text.utf8.decode
-        )
-        .compile
-        .string
-    } yield {
-      val headers = response.headers.headers.groupMap(_.name)(_.value)
-      Some(
-        ApiGatewayProxyStructuredResultV2(
-          response.status.code,
-          (headers - `Set-Cookie`.name).map {
-            case (name, values) =>
-              name -> values.mkString(",")
-          },
-          responseBody,
-          isBase64Encoded,
-          headers.getOrElse(`Set-Cookie`.name, Nil)
-        )
-      )
-    }
-
-  private[http4s] def decodeEvent[F[_]: Concurrent](
-      event: ApiGatewayProxyEventV2): F[Request[F]] = for {
-    method <- Method.fromString(event.requestContext.http.method).liftTo[F]
-    uri <- Uri.fromString(event.rawPath + "?" + event.rawQueryString).liftTo[F]
-    headers = {
-      val builder = List.newBuilder[Header.Raw]
-
-      event.headers.foreachEntry(builder += Header.Raw(_, _))
-      event.cookies.filter(_.nonEmpty).foreach { cs =>
-        builder += Header.Raw(Cookie.name, cs.mkString("; "))
-      }
-
-      Headers(builder.result())
-    }
-    readBody =
-      if (event.isBase64Encoded)
-        fs2.text.base64.decode[F]
-      else
-        fs2.text.utf8.encode[F]
-  } yield Request(
-    method,
-    uri,
-    headers = headers,
-    body = Stream.fromOption[F](event.body).through(readBody)
-  )
+    ApiGatewayProxyHandlerV2(app)
 }
