+++ lambda/shared/src/main/scala-3/feral/lambda/invocations.scala
@@ -19,7 +19,7 @@ package feral.lambda
 import events._
 
 type ApiGatewayProxyInvocation[F[_]] = Invocation[F, ApiGatewayProxyEvent]
-type ApiGatewayProxyV2Invocation[F[_]] = Invocation[F, ApiGatewayProxyEventV2]
+type ApiGatewayProxyInvocationV2[F[_]] = Invocation[F, ApiGatewayProxyEventV2]
 type DynamoDbStreamInvocation[F[_]] = Invocation[F, DynamoDbStreamEvent]
 type S3Invocation[F[_]] = Invocation[F, S3Event]
 type S3BatchInvocation[F[_]] = Invocation[F, S3BatchEvent]
@@ -31,8 +31,8 @@ type LambdaEnv[F[_], Event] = Invocation[F, Event]
 @deprecated("Renamed to Invocation", "0.3.0")
 val LambdaEnv = Invocation
 
-@deprecated("Renamed to ApiGatewayProxyV2Invocation", "0.3.0")
-type ApiGatewayProxyLambdaEnv[F[_]] = ApiGatewayProxyV2Invocation[F]
+@deprecated("Renamed to ApiGatewayProxyInvocationV2", "0.3.0")
+type ApiGatewayProxyLambdaEnv[F[_]] = ApiGatewayProxyInvocationV2[F]
 @deprecated("Renamed to DynamoDbStreamInvocation", "0.3.0")
 type DynamoDbStreamLambdaEnv[F[_]] = DynamoDbStreamInvocation[F]
 @deprecated(
