+++ scalafix/input/src/main/scala/example/V0_3_0Rewrites.scala
@@ -3,6 +3,7 @@
 package example
 
 // format: off
+import cats.effect.Concurrent
 import feral.lambda.LambdaEnv
 import feral.lambda.ApiGatewayProxyLambdaEnv
 import feral.lambda.DynamoDbStreamLambdaEnv
@@ -11,6 +12,9 @@ import feral.lambda.SnsLambdaEnv
 import feral.lambda.SqsLambdaEnv
 import feral.lambda.events.APIGatewayProxyRequestEvent
 import feral.lambda.events.APIGatewayProxyResponseEvent
+import feral.lambda.events.ApiGatewayProxyStructuredResultV2
+import feral.lambda.http4s.ApiGatewayProxyHandler
+import org.http4s.HttpApp
 // format: on
 
 class Foo[F[_], E] {
@@ -20,7 +24,10 @@ class Foo[F[_], E] {
 }
 
 object Handlers {
-  def handler1[F[_]](implicit env: ApiGatewayProxyLambdaEnv[F]): Unit = ???
+  def handler1[F[_]: Concurrent](
+      implicit env: ApiGatewayProxyLambdaEnv[F]
+  ): F[Option[ApiGatewayProxyStructuredResultV2]] =
+    ApiGatewayProxyHandler.httpApp(HttpApp.notFound)
   def handler2[F[_]](implicit env: DynamoDbStreamLambdaEnv[F]): Unit = ???
   def handler3[F[_]](implicit env: S3BatchLambdaEnv[F]): Unit = ???
   def handler4[F[_]](implicit env: SnsLambdaEnv[F]): Unit = ???
